+++ docs/ReleaseNotes.md
@@ -12,7 +12,6 @@ The included licenses apply to the following files:
 
 | License file | Applies to |
 |---|---|
-|LICENSE-MS.txt     |dxil.dll (if included in package)|
 |LICENSE-MIT.txt    |d3d12shader.h|
 |LICENSE-LLVM.txt   |all other files|
 
@@ -23,6 +22,8 @@ The included licenses apply to the following files:
 Place release notes for the upcoming release below this line and remove this line upon naming this release.
 
 - The incomplete WaveMatrix implementation has been removed.
+- DXIL Validator Hash is open sourced.
+- DXIL container validation for PSV0 part allows any content ordering inside string and semantic index tables.
 
 ### Version 1.8.2407